samples: usb: make USB DFU next sample default
Move legacy USB DFU sample to legacy directory and make USB DFU next
sample default.
