--- samples/subsys/usb/dfu-next/prj.conf
@@ -1,10 +1,0 @@
-CONFIG_USB_DEVICE_STACK_NEXT=y
-CONFIG_USBD_DFU=y
-
-CONFIG_LOG=y
-CONFIG_USBD_LOG_LEVEL_WRN=y
-CONFIG_USBD_DFU_LOG_LEVEL_WRN=y
-CONFIG_UDC_DRIVER_LOG_LEVEL_WRN=y
-CONFIG_SAMPLE_USBD_PID=0x0005
-
-CONFIG_DISK_ACCESS=y
