+++ samples/subsys/usb/dfu/app.overlay
