+++ samples/subsys/usb/dfu/prj.conf
@@ -1,16 +1,10 @@
-CONFIG_STDOUT_CONSOLE=y
-CONFIG_GPIO=y
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_USB_DEVICE_PRODUCT="Zephyr DFU sample"
-CONFIG_USB_DEVICE_PID=0x0005
-CONFIG_USB_DFU_CLASS=y
-CONFIG_USB_DFU_ENABLE_UPLOAD=y
-CONFIG_FLASH=y
-CONFIG_FLASH_MAP=y
-CONFIG_STREAM_FLASH=y
-CONFIG_IMG_MANAGER=y
+CONFIG_USB_DEVICE_STACK_NEXT=y
+CONFIG_USBD_DFU=y
+
 CONFIG_LOG=y
-CONFIG_USB_DRIVER_LOG_LEVEL_ERR=y
-CONFIG_USB_DEVICE_LOG_LEVEL_ERR=y
-CONFIG_BOOTLOADER_MCUBOOT=y
-CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
+CONFIG_USBD_LOG_LEVEL_WRN=y
+CONFIG_USBD_DFU_LOG_LEVEL_WRN=y
+CONFIG_UDC_DRIVER_LOG_LEVEL_WRN=y
+CONFIG_SAMPLE_USBD_PID=0x0005
+
+CONFIG_DISK_ACCESS=y
