+++ samples/subsys/usb/dfu/sample.yaml
@@ -1,45 +1,30 @@
 sample:
   name: USB DFU sample
-
 common:
-  build_only: true
-  arch_exclude: posix
-  platform_exclude:
-    - mimxrt1010_evk
-    - mimxrt1020_evk
-    - mimxrt1015_evk
-    - mimxrt1060_evk@A/mimxrt1062/qspi
-    - mimxrt1060_evk@B/mimxrt1062/qspi
-    - mimxrt1060_evk@C/mimxrt1062/qspi
-    - sam4l_ek
-    - same54_xpro
-    - samr21_xpro
-    - mimxrt1050_evk/mimxrt1052/hyperflash
-    - mimxrt1050_evk/mimxrt1052/qspi
-    - mimxrt1060_evk/mimxrt1062/hyperflash
-    - mimxrt1180_evk/mimxrt1189/cm7
-    - mimxrt700_evk/mimxrt798s/cm33_cpu0
-    - nucleo_f207zg
-    - teensy40
-    - teensy41
-    - b_u585i_iot02a
-    - frdm_kl25z
-    - frdm_mcxc242
-    - frdm_mcxc444
-    - frdm_mcxa153
-    - frdm_mcxa156
-    - lpcxpresso55s69/lpc55s69/cpu0
-    - stm32l562e_dk/stm32l562xx/ns
-  depends_on: usb_device
-  filter: dt_label_with_parent_compat_enabled("slot0_partition", "fixed-partitions")
-    and dt_label_with_parent_compat_enabled("slot1_partition", "fixed-partitions")
-    and dt_chosen_enabled("zephyr,flash-controller") and CONFIG_FLASH_HAS_DRIVER_ENABLED
-  integration_platforms:
-    - nrf52840dk/nrf52840
-    - frdm_k64f
+  min_ram: 64
+  depends_on:
+    - usbd
+  harness: console
+  harness_config:
+    type: one_line
+    regex:
+      - "USB DFU sample is initialized"
 tests:
-  sample.usb.dfu:
+  sample.usbd.dfu:
+    integration_platforms:
+      - nrf52840dk/nrf52840
+      - nrf54h20dk/nrf54h20/cpuapp
+      - frdm_k64f
+      - stm32f723e_disco
+      - nucleo_f413zh
+      - mimxrt685_evk/mimxrt685s/cm33
+      - mimxrt1060_evk/mimxrt1062/qspi
     tags: usb
-  sample.usb.dfu.permanent.download:
+  sample.usbd.dfu-flash:
+    platform_allow:
+      - nrf52840dk/nrf52840
+      - frdm_k64f
+    min_flash: 1024
     tags: usb
-    extra_args: EXTRA_CONF_FILE=overlay-permanent-download.conf
+    extra_configs:
+      - CONFIG_APP_USB_DFU_USE_FLASH_BACKEND=y
