+++ samples/subsys/usb/legacy/dfu/overlay-permanent-download.conf
